!!!**READ DESC**!!!
changed some colors, changed the font of “welcome” on index page,
tested out an idea that we could use for the website on the
“rollercoasters” page (??what do you guys think??). i was trying to
figure out how to make the nav bar stick to the top to the page when
you scroll up to it, its proving to be quite difficult. if you
know/could figure that out it’d be great! if not i will.